+++ Source/CartLaunchCompanion.Core/Updating/OfficialUpdateTrust.cs
@@ -1,0 +1,13 @@
+namespace CartLaunchCompanion.Core.Updating;
+
+public static class OfficialUpdateTrust
+{
+    public const string Repository = "Uplinkpro/CartLaunchCompanion";
+    public const string KeyId = "uplinkpro-release-1";
+    public const string PublicKeyPem = """
+        -----BEGIN PUBLIC KEY-----
+        MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEwSnjXV9pP/RuRH98cJXSq8Lw6332
+        wssHkj7Ar7CRdv4eXonx06il4+hqLG19fe0R0Qz5XfuYZqWR5/J4qcTR6A==
+        -----END PUBLIC KEY-----
+        """;
+}