+++ Source/CartLaunchCompanion.Updater/UpdateTrustAnchor.cs
@@ -1,12 +1,9 @@
+using CartLaunchCompanion.Core.Updating;
+
 namespace CartLaunchCompanion.Updater;
 
 internal static class UpdateTrustAnchor
 {
-    public const string OfficialKeyId = "uplinkpro-release-1";
-    public const string OfficialPublicKeyPem = """
-        -----BEGIN PUBLIC KEY-----
-        MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEwSnjXV9pP/RuRH98cJXSq8Lw6332
-        wssHkj7Ar7CRdv4eXonx06il4+hqLG19fe0R0Qz5XfuYZqWR5/J4qcTR6A==
-        -----END PUBLIC KEY-----
-        """;
+    public const string OfficialKeyId = OfficialUpdateTrust.KeyId;
+    public const string OfficialPublicKeyPem = OfficialUpdateTrust.PublicKeyPem;
 }